+++ logprep/connector/confluent_kafka/input.py
@@ -344,7 +344,7 @@ class ConfluentKafkaInput(Input):
             the error that occurred
         """
         self.metrics.number_of_errors += 1
-        logger.error("%s: %s", self.describe(), error)
+        logger.error("%s: %s", self.description, error)
 
     def _stats_callback(self, stats_raw: str) -> None:
         """Callback for statistics data. This callback is triggered by poll()
@@ -412,7 +412,7 @@ class ConfluentKafkaInput(Input):
             }
             self.metrics.committed_offsets.add_with_labels(offset, labels)
 
-    def describe(self) -> str:
+    def _describe(self) -> str:
         """Get name of Kafka endpoint and bootstrap servers.
 
         Returns
@@ -420,8 +420,9 @@ class ConfluentKafkaInput(Input):
         kafka : str
             Description of the ConfluentKafkaInput connector.
         """
-        base_description = super().describe()
-        return f"{base_description} - Kafka Input: {self.config.kafka_config['bootstrap.servers']}"
+        return (
+            f"{super()._describe()} - Kafka Input: {self.config.kafka_config['bootstrap.servers']}"
+        )
 
     def _get_raw_event(self, timeout: float) -> bytes | None:
         """Get next raw Message from Kafka.