+++ tests/unit/ng/connector/test_opensearch_output.py
@@ -103,7 +103,7 @@ class TestOpenSearchOutput(BaseOutputTestCase[OpensearchOutput]):
 
     async def test_describe_returns_output(self):
         assert (
-            self.object.describe()
+            self.object.description
             == "OpensearchOutput (Test Instance Name) - Opensearch Output: ['localhost:9200']"
         )